refactor: core/utils/ -> core/{validation,worker,util}/ + presentation/{util->common}/
