+++ app/src/main/java/com/drs/auralife/core/util/Notification.kt
@@ -1,4 +1,4 @@
-package com.drs.auralife.core.utils
+package com.drs.auralife.core.util
 
 import android.content.Context
 import org.json.JSONArray
