+++ app/src/main/java/com/drs/auralife/data/repository/AvatarRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.drs.auralife.data.repository
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import com.drs.auralife.core.utils.ImageEncoderDecoder
+import com.drs.auralife.core.util.ImageEncoderDecoder
 import com.drs.auralife.data.remote.firebase.Authentication
 import com.drs.auralife.data.remote.firebase.AvatarDataSource as FirebaseAvatarRepository
 import com.drs.auralife.domain.repository.AvatarRepository
