+++ app/src/main/java/com/drs/auralife/presentation/common/MyAppGlideModule.kt
@@ -1,4 +1,4 @@
-package com.drs.auralife.core.utils
+package com.drs.auralife.presentation.common
 
 import android.app.Activity
 import android.content.Context
