+++ app/src/main/java/com/drs/auralife/presentation/explore/adapter/CategoryFilmAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.drs.auralife.R
-import com.drs.auralife.core.utils.MyAppGlideModule
+import com.drs.auralife.presentation.common.MyAppGlideModule
 import com.drs.auralife.domain.model.Film
 
 class CategoryFilmAdapter(
