+++ app/src/main/java/com/drs/auralife/presentation/payment/PaymentFragment.kt
@@ -14,7 +14,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
-import com.drs.auralife.presentation.util.launchAndRepeatWithViewLifecycle
+import com.drs.auralife.presentation.common.launchAndRepeatWithViewLifecycle
 import com.drs.auralife.R
 import com.drs.auralife.databinding.ActivityPaymentBinding
 import com.drs.auralife.domain.model.PaymentItem
